Don't fail the task when adding a directory that doesn't exist to PATH
