+++ Tasks/UsePythonVersion/Strings/resources.resjson/en-US/resources.resjson
@@ -9,5 +9,6 @@
   "loc.input.help.addToPath": "Whether to prepend the retrieved Python version to the PATH environment variable to make it available in subsequent tasks or scripts without using the output variable.",
   "loc.messages.ListAvailableVersions": "Available versions:",
   "loc.messages.PlatformNotRecognized": "Platform not recognized",
+  "loc.messages.PrependPath": "Prepending PATH environment variable with directory: %s",
   "loc.messages.VersionNotFound": "Version spec %s did not match any version in the tool cache."
 }
